refactor(SPV-1173): move keysgen one level upper.
